==> 0a - scraping_google_books.ipynb <==
# Prepare notebook

from bs4 import BeautifulSoup
import googlesearch
import isbnlib
import json
import pandas as pd
import urllib
import wikipedia



API_key = ''

# Load data set, clean, & scrape web for additional information

books = pd.read_csv('./goodreads_library_export.csv')

books = books[['ISBN13', 'Title', 'Author', 'My Rating','Publisher',
       'Number of Pages', 'Year Published', 'Original Publication Year',
       'Date Added', 'Bookshelves', 'Exclusive Shelf',
       'Private Notes']]

books = books.loc[~(books['Exclusive Shelf'] == 'to-read')]
del books['Exclusive Shelf']

books['Date Added'] = pd.to_datetime(books['Date Added'])

books['recommended'] = books.Bookshelves.fillna(False).astype(bool)
del books['Bookshelves']

books['ISBN13'] = books['ISBN13'].apply(lambda s: s.replace('=','').replace('"', '').strip())

books.to_csv('./goodreads_library_export_cleaned.csv')

scraped_data = {}

for index, row in books.fillna('').iterrows():
    scraped_data[index] = {'isbn': None,
                           'categories': None,
                           'description': None}

    # Pull the Google Books information 
    try:
        base_api_link = "https://www.googleapis.com/books/v1/volumes?q=isbn:<isbn>&key="+API_key
        with urllib.request.urlopen(base_api_link.replace('<isbn>', row['ISBN13'])) as f:
            text = f.read()

        decoded_text = text.decode("utf-8")
        google_books_data = json.loads(decoded_text) # deserializes decoded_text to a Python object
        for identifier in google_books_data['items'][0]['volumeInfo']['industryIdentifiers']:
            if identifier['type'] == 'ISBN_13':
                scraped_data[index]['isbn'] = identifier['identifier']
                break
        scraped_data[index]['categories'] = google_books_data['items'][0]['volumeInfo']['categories']
        scraped_data[index]['description'] = google_books_data['items'][0]['volumeInfo']['description']
    except:
        try:
            base_api_link = "https://www.googleapis.com/books/v1/volumes?q=intitle:<title>+inauthor:<author>&key="+API_key
            with urllib.request.urlopen(base_api_link.replace('<title>', row['Title'].replace(' ', '+'))
                                                      .replace('<author>', row['Author'].replace(' ', '+'))) as f:
                text = f.read()

            decoded_text = text.decode("utf-8")
            google_books_data = json.loads(decoded_text) # deserializes decoded_text to a Python object
            for identifier in google_books_data['items'][0]['volumeInfo']['industryIdentifiers']:
                if identifier['type'] == 'ISBN_13':
                    scraped_data[index]['isbn'] = identifier['identifier']
                    break
            scraped_data[index]['categories'] = google_books_data['items'][0]['volumeInfo']['categories']
            scraped_data[index]['description'] = google_books_data['items'][0]['volumeInfo']['description']
        except:
            print('Unable to collect Google Books data for %s'%row['Title'])

books = books[[ 'Title', 'Author', 'Publisher', 'My Rating','Number of Pages', 'Year Published', 
                    'Original Publication Year', 'Date Added', 'recommended']].join(pd.DataFrame(scraped_data).T)

books.count()

books.to_pickle('20191128_google_book_data.pkl')
