==> 0b - scraping_goodreads.ipynb <==
# Prepare notebook

from ast import literal_eval
from bs4 import BeautifulSoup
import json
import googlesearch
from googletrans import Translator
import pandas as pd
import urllib

# Load data set, clean, & scrape web for additional information

books = pd.read_csv('./goodreads_library_export.csv')

books = books[['ISBN13', 'Title', 'Author', 'My Rating','Publisher',
       'Number of Pages', 'Year Published', 'Original Publication Year',
       'Date Added', 'Bookshelves', 'Exclusive Shelf',
       'Private Notes']]

books = books.loc[~(books['Exclusive Shelf'] == 'to-read')]
del books['Exclusive Shelf']

books['Date Added'] = pd.to_datetime(books['Date Added'])

books['recommended'] = books.Bookshelves.fillna(False).astype(bool)
del books['Bookshelves']

books['ISBN13'] = books['ISBN13'].apply(lambda s: s.replace('=','').replace('"', '').strip())

translator = Translator()
scraped_data = {}

for index, row in books.fillna('').iterrows():
    scraped_data[index] = {'goodreads_link': None,
                           'goodreads_shelves': [],
                           'goodreads_description': None,
                           'goodreads_author_bio': None}
    
    
    # Pull the Goodreads webpage
    for result in googlesearch.search('Goodreads '+ row['Title']+' '+row['Author'], 
                                      start = 0, stop =10):
        if ('goodreads' in result) & ('book/show' in result):
            print(result)
            scraped_data[index]['goodreads_link'] = result
            try:
                http_query_result= urllib.request.urlopen(result)
            except urllib.error.HTTPError:
                print('Error reading http')
                break
            else: 
                soup = BeautifulSoup(http_query_result, 'html.parser')
                for shelf in soup.select("a[title*=shelved]"):
                    scraped_data[index]['goodreads_shelves'] += [(shelf.attrs["href"]).split('=')[1]]
                #if len(scraped_data[index]['goodreads_shelves']) == 0:
                    #scraped_data[index]['goodreads_shelves'] = literal_eval(soup.text.split('googletag.pubads().setTargeting("shelf", ')[1].split(');')[0])
                if len(scraped_data[index]['goodreads_shelves']) == 0:
                    print('  Missing shelves')

                try:
                    description = soup.find("div", {'id':"description"}).text
                    language = translator.detect(description).lang
                    if language != 'en':
                        description = translator.translate(description, dest='en', src = language).text
                    scraped_data[index]['goodreads_description'] = description        
                except:
                    print('  Missing description') 
                    
                    
                try:    
                    scraped_data[index]['goodreads_author_bio'] = soup.find("div", {'class':'bookAuthorProfile__about'}).text
                except:
                    print('  No author bio available')
                break

books = books[[ 'Title', 'Author', 'Publisher', 'My Rating','Number of Pages', 'Year Published', 
                    'Original Publication Year', 'Date Added', 'recommended']].join(pd.DataFrame(scraped_data).T)

books.count()

books.to_pickle('20191128_goodreads_book_data.pkl')

# Manually short shelves

shelves = pd.get_dummies(books.goodreads_shelves.apply(pd.Series).stack()).sum(level=0)
shelves = shelves.astype(bool)
shelves = shelves.reindex(books.index)

category_mappings = {'literature_fiction':[ 'arthurian',
                                             'british-literature',
                                             'classic-literature',
                                             'classics',
                                             'fairy-tales',
                                             'fantasy',
                                             'fiction',
                                             'folk-tales',
                                             'folklore',
                                             'historical-fiction',
                                             'italian-literature',
                                             'japanese-literature',
                                             'literary-fiction',
                                             'literature',
                                             'military-fiction',
                                             'mythology',
                                             'novels',
                                             'russian-literature',
                                             'science-fiction',
                                             'short-stories',
                                             'spanish-literature',
                                             'urban-fantasy'],
                    'history': ['10th-century',
                                '19th-century',
                                 'american',
                                 'american-history',
                                 'ancient-history',
                                 'art-history',
                                 'church-history',
                                 'civil-war',
                                 'european-history',
                                 'egyptology',
                                 'food-history',
                                 'gothic',
                                 'historical',
                                 'history',
                                 'history-of-science',
                                 'holocaust',
                                 'medieval',
                                 'medieval-history',
                                 'microhistory',
                                 'post-colonial',
                                 'prehistory',
                                 'world-history',
                                 'world-war-ii'],
                    'social_studies': [ 'activism',
                                         'african-american',
                                         'anthropology',
                                         'business',
                                         'communication',
                                         'crime',
                                         'cultural',
                                         'disability',
                                         'economics',
                                         'education',
                                         'entrepreneurship',
                                         'espionage',
                                         'feminism',
                                         'finance',
                                         'futurism',
                                         'gay',
                                         'gender',
                                         'gender-studies',
                                         'journalism',
                                         'language',
                                         'leadership',
                                         'lgbt',
                                         'linguistics',
                                         'management',
                                         'money',
                                         'parenting',
                                         'politics',
                                         'productivity',
                                         'queer',
                                         'race',
                                         'social-justice',
                                         'society',
                                         'sociology',
                                         'true-crime',
                                         'war',
                                         'witchcraft',
                                         'womens'],
                    'technology': [ 'artificial-intelligence',
                                     'computer-science',
                                     'computers',
                                     'hackers',
                                     'internet',
                                     'mathematics',
                                     'physics',
                                     'technology'],
                    'culture': [ 'adventure',
                                 'architecture',
                                 'art',
                                 'basketball',
                                 'books-about-books',
                                 'comedy',
                                 'cookbooks',
                                 'cooking',
                                 'crafts',
                                 'design',
                                 'drama',
                                 'essays',
                                 'fashion',
                                 'food',
                                 'gardening',
                                 'hip-hop',
                                 'horror',
                                 'humor',
                                 'jazz',
                                 'magic',
                                 'maps',
                                 'music',
                                 'mystery',
                                 'photography',
                                 'plays',
                                 'poetry',
                                 'sports',
                                 'tea',
                                 'theatre',
                                 'thriller',
                                 'tragedy',
                                 'travel',
                                 'writing'],
                    'natural_sciences': [ 'agriculture',
                                         'animals',
                                         'archaeology',
                                         'astronomy',
                                         'biology',
                                         'birds',
                                         'cartography',
                                         'cats',
                                         'chemistry',
                                         'conservation',
                                         'dinosaurs',
                                         'dogs',
                                         'earth-sciences',
                                         'ecology',
                                         'environment',
                                         'evolution',
                                         'geography',
                                         'geology',
                                         'natural-history',
                                         'nature',
                                         'palaeontology',
                                         'plants',
                                         'research',
                                         'science',
                                         'science-nature',
                                         'space',
                                         'wildlife'],
                    'philosophy': [ 'atheism',
                                     'buddhism',
                                     'christianity',
                                     'inspirational',
                                     'islam',
                                     'judaism',
                                     'personal-development',
                                     'philosophy',
                                     'religion',
                                     'self-help',
                                     'spirituality',
                                     'zen'],
                    'biography_memoir': [ 'autobiography',
                                         'biography',
                                         'biography-memoir',
                                         'memoir'],
                    'medicine':  ['brain',
                                 'disease',
                                 'fitness',
                                 'genetics',
                                 'health',
                                 'health-care',
                                 'medical',
                                 'medicine',
                                 'mental-health',
                                 'neuroscience',
                                 'nutrition',
                                 'psychology']
                    }

categories = pd.DataFrame(index = shelves.index)
for k,v in category_mappings.items():
    categories[k] = (shelves[v].sum(axis =1) > 0)

categories.sum()

shelves.to_pickle('20191128_goodreads_book_shelves.pkl')

categories.to_pickle('20191128_goodreads_book_categories.pkl')
json.dump(category_mappings, open("category_mappings.json","w"))
