==> 0c - additional preprocessing.ipynb <==
# Prepare notebook

import json 
import pandas as pd

# Import data

books = pd.read_pickle('2019112_goodreads_book_data.pkl')

# Initial exploration of the data

shelves = pd.get_dummies(books.goodreads_shelves.apply(pd.Series).stack()).sum(level=0)

shelves = shelves.astype(bool)

shelves.sum().sort_values(ascending = False).iloc[0:30]

shelves = shelves.reindex(books.index)

academic_field_mappings  = { 'humanities': ['archaeology','art','architecture','cartography','history','music', 
                                'library-science', 'linguistics'],
                 
                      'social_sciences': ['business','economics','education','finance','psychology','geography',
                                          'political-science','sociology','social-science', 'journalism', 
                                          'management'],

                      'physical_sciences':['anthropology','astronomy',
                                        'biology', 'natural-history',
                                        'chemistry','computer-science','ecology',
                                        'earth-sciences','engineering',
                                        'geology','medicine',  
                                         'neuroscience',
                                        'palaeontology',   'mathematics','physics'],
                     }



category_mappings = {'history': ['history','19th-century','archaeology','ancient','american-history','art-history','ancient-history',
                             'classical-studies','civil-war','egyptology','european-history','food-history','historical', 
                             'history-of-science','holocaust', 'medieval', 'medieval-history', 
                             'medievalism', 'microhistory','military-history', 'prehistory', 'roman', 'victorian',
                             'war', 'world-history', 'world-war-ii','palaeontology'],
                 
                 'sociology': ['activism', 'african-american','communication','crime','disability',
                              'family','feminism','gay','gender','gender-studies','health-care', 'social-justice',
                               'relationships', 'leadership', 'lgbt', 'politics', 'parenting', 'money', 'productivity',
                               'queer','race', 'social-issues',  'society',  'witchcraft',
                               'womens','business','economics','education','finance','psychology','geography',
                                'political-science','sociology','social-science', 'journalism', 
                                'management'],
                        
                 'sciences':['ecology','earth-sciences','biology', 'natural-history','anthropology','animals','birds','cats','conservation',
                                     'dinosaurs','dogs','science-nature','geology',
                                     'environment','evolution', 'genetics', 'nature', 'ornithology',
                                     'outdoors', 'plants', 'space', 'wildlife', 'science','astronomy',
                                   'chemistry','mathematics','physics'],
                 
                 'technology':['computer-science','artificial-intelligence','computers',
                               'hackers', 'internet', 'technology', 'theory','engineering'],
                 
                 'philosophy':['atheism','buddhism','christianity','eastern-philosophy','inspirational','islam',
                               'judaism','occult', 'philosophy', 'spirituality', 'religion', 'taoism'],
                
                 
                 'culture':['art','architecture','cartography','alcohol','basketball', 'classical-music','cultural','book-club','crafts',
                            'design','espionage','fashion','gardening', 'jazz','music', 
                             'photography', 'sports', 'theatre',  'language', 'librarianship',
                            'writing', 'magic','maps','witches','plays', 'poetry', 'classics','library-science', 'linguistics'],
                 
                 'literatures':['literature', 'british-literature','classic-literature',
                                    'english-literature', 'french-literature', 'italian-literature', 'japanese-literature',
                                     'russian-literature', 'spanish-literature'],
                 
                  'other':['comedy', 'drama','fantasy','horror', 'humor',  'mystery',
                            'romance', 'novels', 'science-fiction', 'short-stories', 'thriller', 'tragedy',
                                'true-crime','books-about-books','collections','essays'],
                 
                  'mythology': ['arthurian','fairy-tales', 'folk-tales','folklore', 'mythology'],
                 
                 'biography':['autobiography','biography','biography-memoir','diary', 'memoir','travelogue'],
                 
                
                 'travel':['africa','adventure','american','asia','australia','canada','china','egypt', 
                           'france','greece','india', 'ireland', 'travel', 'italy', 'japan', 'new-york', 'pakistan',
                           'russia', 'spain', 'the-united-states-of-america', 'united-states', 'western-africa'],
                 
                 'culinary': ['cookbooks','cooking','culinary','food','food-and-drink', 'foodie', 'tea'],
                 
                 'health':['medicine', 'neuroscience','brain','health','fitness', 'medical',  'mental-health', 'nutrition', 'personal-development', 'self-help']
                }

 

set(shelves.columns.tolist()).difference(set([item for sublist in genre_mapping.values() for item in sublist]))


academic_fields = pd.DataFrame(index = shelves.index)
for k,v in academic_field_mappings.items():
    academic_fields[k] = (shelves[v].sum(axis =1) > 0)

categories = pd.DataFrame(index = shelves.index)
for k,v in category_mappings.items():
    categories[k] = (shelves[v].sum(axis =1) > 0)

categories.sum()

academic_fields.sum()

shelves.to_pickle('2019112_goodreads_book_shelves.pkl')

categories.to_pickle('2019112_goodreads_book_categories.pkl')
json.dump(category_mappings, open("category_mappings.json","w"))

academic_fields.to_pickle('2019112_goodreads_book_academic_fields.pkl')
json.dump(academic_field_mappings, open("academic_field_mappings.json","w"))
