==> 3a - publisher_normalization.ipynb <==
# Prepare notebook

import numpy as np
import pandas as pd
import re
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

# Load data set, clean, & scrape web for additional information

publishers = pd.read_csv('./goodreads_library_export_cleaned.csv',  index_col = 0)[['Publisher']].fillna('')


publishers['normalized'] = publishers['Publisher'].apply(lambda s: s.split('/')[0].strip())



unique_publishers = publishers['normalized'].drop_duplicates()
unique_publishers  = unique_publishers.reset_index(drop = True)

publisher_counts = publishers['normalized'].value_counts()

vectorizer = CountVectorizer(analyzer = 'word', stop_words = ['press', 'publisher', 'publishing', 'publishers',
                                                               'university', 'group', 'book', 'books'])#analyzer = 'char', ngram_range = (5,5), )
publisher_ngrams = vectorizer.fit_transform(unique_publishers)
#print(vectorizer.get_feature_names())

similarities = np.triu(cosine_similarity(publisher_ngrams))

matches = np.where((similarities > 0.5) & (similarities < 0.9999))

d = {}

for i, m1 in enumerate(matches[0]):
    if m1 not in d.keys():
        d[m1] = [unique_publishers[m1], unique_publishers[matches[1][i]]]    
    else:
        d[m1] += [unique_publishers[matches[1][i]]] 
                

publisher_map = {}
for v in d.values():
    standard_name = v[np.argmax([publisher_counts[p] for p in v])]
    v.remove(standard_name) 
    for p in v:
        publisher_map[p]=standard_name

for k,v in publisher_map.items():
    if v in publisher_map.keys():
        publisher_map[k] = publisher_map[v]

publisher_map = pd.DataFrame.from_dict(publisher_map, orient = 'index')
publisher_map.index.name = 'old_name'
publisher_map = publisher_map.rename(columns = {0:'new_name'})


publishers['index'] = publishers.index
publishers = pd.DataFrame(publishers).merge(publisher_map, left_on = 'normalized', 
                                            right_on = 'old_name', how = 'left')
publishers = publishers.set_index('index')

idx = publishers.new_name.notnull()
publishers.loc[idx, 'normalized'] = publishers.loc[idx, 'new_name']

publishers['stripped'] =publishers['normalized'].apply(lambda s: re.sub("([a-z])([A-Z])","\g<1> \g<2>", s)
                                                       .lower().replace('press', '')
                                                       .replace('publishers', '')
                                                       .replace('publishing', '')
                                                       .replace('publisher', '')
                                                       .replace('publications', '')
                                                       .replace('university', '')
                                                       .replace('square', '')
                                                       .replace('group', '')
                                                       .replace('books', '')
                                                       .replace('book', '')
                                                       .replace('editions', '')
                                                       .replace('audio','')
                                                       .strip()
                                                      )

unique_publishers  = publishers[['normalized','stripped']].reset_index(drop = True)

publisher_counts = publishers['normalized'].value_counts()

vectorizer = CountVectorizer(analyzer = 'char', ngram_range = (4,4), max_df = 0.7)
publisher_ngrams = vectorizer.fit_transform(unique_publishers['stripped'])
#print(vectorizer.get_feature_names())

similarities = np.triu(cosine_similarity(publisher_ngrams))

matches = np.where((similarities > 0.6) & (similarities < 0.9999))

d = {}

for i, m1 in enumerate(matches[0]):
    if m1 not in d.keys():
        d[m1] = [unique_publishers.loc[m1, 'normalized'], unique_publishers.loc[matches[1][i],'normalized']]    
    else:
        d[m1] += [unique_publishers.loc[matches[1][i],'normalized']] 
                

publisher_map = {}
for v in d.values():
    standard_name = v[np.argmax([publisher_counts[p] for p in v])]
    v.remove(standard_name) 
    for p in v:
        publisher_map[p]=standard_name

for k,v in publisher_map.items():
    if v in publisher_map.keys():
        publisher_map[k] = publisher_map[v]

publisher_map = pd.DataFrame.from_dict(publisher_map, orient = 'index')
publisher_map.index.name = 'old_name'
publisher_map = publisher_map.rename(columns = {0:'new_name_2'})


publishers['index'] = publishers.index
publishers = pd.DataFrame(publishers).merge(publisher_map, left_on = 'normalized', 
                                            right_on = 'old_name', how = 'left')
publishers = publishers.set_index('index')

idx = publishers.new_name_2.notnull()
publishers.loc[idx, 'normalized'] = publishers.loc[idx, 'new_name_2']

len(publishers.normalized.unique())

len(publishers.Publisher.unique())

publishers.normalized.value_counts()

output = publishers[['normalized']].rename(columns = {'normalized': 'publisher'})
output.to_pickle('20191128_normalized_publisher.pkl')

publishers
